Move functions to signals.py
