+++ dojo/endpoint/views.py
@@ -1,22 +1,17 @@
 import logging
 from datetime import datetime
 
-from auditlog.models import LogEntry
 from dateutil.relativedelta import relativedelta
 from django.apps import apps
 from django.conf import settings
 from django.contrib import messages
 from django.contrib.admin.utils import NestedObjects
-from django.contrib.contenttypes.models import ContentType
 from django.db import DEFAULT_DB_ALIAS
 from django.db.models import Count, Q, QuerySet
-from django.db.models.signals import post_delete
-from django.dispatch import receiver
 from django.http import HttpResponseRedirect
 from django.shortcuts import get_object_or_404, render
 from django.urls import reverse
 from django.utils import timezone
-from django.utils.translation import gettext as _
 
 from dojo.authorization.authorization import user_has_permission_or_403
 from dojo.authorization.authorization_decorators import user_is_authorized
@@ -26,7 +21,6 @@ from dojo.endpoint.utils import clean_hosts_run, endpoint_meta_import
 from dojo.filters import EndpointFilter, EndpointFilterWithoutObjectLookups
 from dojo.forms import AddEndpointForm, DeleteEndpointForm, DojoMetaDataForm, EditEndpointForm, ImportEndpointMetaForm
 from dojo.models import DojoMeta, Endpoint, Endpoint_Status, Finding, Product
-from dojo.notifications.helper import create_notification
 from dojo.utils import (
     Product_Tab,
     add_breadcrumb,
@@ -515,23 +509,3 @@ def import_endpoint_meta(request, pid):
         'product_tab': product_tab,
         'form': form,
     })
-
-
-@receiver(post_delete, sender=Endpoint)
-def endpoint_post_delete(sender, instance, using, origin, **kwargs):
-    if instance == origin:
-        if get_system_setting('enable_auditlog'):
-            le = LogEntry.objects.get(
-                    action=LogEntry.Action.DELETE,
-                    content_type=ContentType.objects.get(app_label='dojo', model='endpoint'),
-                    object_id=instance.id
-            )
-            description = _('The endpoint "%(name)s" was deleted by %(user)s') % {
-                                'name': str(instance), 'user': le.actor}
-        else:
-            description = _('The endpoint "%(name)s" was deleted') % {'name': str(instance)}
-        create_notification(event='endpoint_deleted',  # template does not exists, it will default to "other" but this event name needs to stay because of unit testing
-                            title=_('Deletion of %(name)s') % {'name': str(instance)},
-                            description=description,
-                            url=reverse('endpoint'),
-                            icon="exclamation-triangle")
