+++ dojo/product/views.py
@@ -6,19 +6,15 @@ from collections import OrderedDict
 from datetime import date, datetime, timedelta
 from math import ceil
 
-from auditlog.models import LogEntry
 from dateutil.relativedelta import relativedelta
 from django.contrib import messages
 from django.contrib.admin.utils import NestedObjects
-from django.contrib.contenttypes.models import ContentType
 from django.contrib.postgres.aggregates import StringAgg
 from django.core.exceptions import PermissionDenied, ValidationError
 from django.db import DEFAULT_DB_ALIAS, connection
 from django.db.models import Count, F, Max, OuterRef, Prefetch, Q, Subquery, Sum
 from django.db.models.expressions import Value
 from django.db.models.query import QuerySet
-from django.db.models.signals import post_delete, post_save
-from django.dispatch import receiver
 from django.http import Http404, HttpRequest, HttpResponseRedirect, JsonResponse
 from django.shortcuts import get_object_or_404, render
 from django.urls import reverse
@@ -95,7 +91,6 @@ from dojo.models import (
     Test,
     Test_Type,
 )
-from dojo.notifications.helper import create_notification
 from dojo.product.queries import (
     get_authorized_groups_for_product,
     get_authorized_members_for_product,
@@ -1988,31 +1983,3 @@ def add_product_group(request, pid):
         'form': group_form,
         'product_tab': product_tab,
     })
-
-
-@receiver(post_save, sender=Product)
-def product_post_save(sender, instance, created, **kwargs):
-    if created:
-        create_notification(event='product_added',
-                            title=instance.name,
-                            product=instance,
-                            url=reverse('view_product', args=(instance.id,)))
-
-
-@receiver(post_delete, sender=Product)
-def product_post_delete(sender, instance, **kwargs):
-    if get_system_setting('enable_auditlog'):
-        le = LogEntry.objects.get(
-                action=LogEntry.Action.DELETE,
-                content_type=ContentType.objects.get(app_label='dojo', model='product'),
-                object_id=instance.id
-        )
-        description = _('The product "%(name)s" was deleted by %(user)s') % {
-                            'name': instance.name, 'user': le.actor}
-    else:
-        description = _('The product "%(name)s" was deleted') % {'name': instance.name}
-    create_notification(event='product_deleted',  # template does not exists, it will default to "other" but this event name needs to stay because of unit testing
-                        title=_('Deletion of %(name)s') % {'name': instance.name},
-                        description=description,
-                        url=reverse('product'),
-                        icon="exclamation-triangle")
