+++ dojo/test/signals.py
@@ -1,15 +1,40 @@
 import contextlib
-import logging
 
-from django.db.models import signals
+from auditlog.models import LogEntry
+from django.contrib.contenttypes.models import ContentType
+from django.db.models.signals import post_delete, pre_save
 from django.dispatch import receiver
+from django.urls import reverse
+from django.utils.translation import gettext as _
 
 from dojo.models import Finding, Test
+from dojo.notifications.helper import create_notification
+from dojo.utils import get_system_setting
 
-logger = logging.getLogger(__name__)
+
+@receiver(post_delete, sender=Test)
+def test_post_delete(sender, instance, using, origin, **kwargs):
+    if instance == origin:
+        if get_system_setting('enable_auditlog'):
+            le = LogEntry.objects.get(
+                    action=LogEntry.Action.DELETE,
+                    content_type=ContentType.objects.get(app_label='dojo', model='test'),
+                    object_id=instance.id
+            )
+            description = _('The test "%(name)s" was deleted by %(user)s') % {
+                                'name': str(instance), 'user': le.actor}
+        else:
+            description = _('The test "%(name)s" was deleted') % {'name': str(instance)}
+        create_notification(event='test_deleted',  # template does not exists, it will default to "other" but this event name needs to stay because of unit testing
+                            title=_('Deletion of %(name)s') % {'name': str(instance)},
+                            description=description,
+                            product=instance.engagement.product,
+                            url=reverse('view_engagement', args=(instance.engagement.id, )),
+                            recipients=[instance.engagement.lead],
+                            icon="exclamation-triangle")
 
 
-@receiver(signals.pre_save, sender=Test)
+@receiver(pre_save, sender=Test)
 def update_found_by_for_findings(sender, instance, **kwargs):
     with contextlib.suppress(sender.DoesNotExist):
         obj = sender.objects.get(pk=instance.pk)
